requirements.txt.txt
tensorflow==2.6.0
streamlit==1.0.0
tensorflow_hub==0.12.0
numpy==1.19.5
matplotlib==3.2.2
Pillow==7.1.2